rework accept function
The previous design couldn't for example accept only valid chess moves
because it didn't have access to the state.

The new is more ergonomic, as it allows returning just a part of the input.
That is what is needed in the typical case where input is an enum and
only some of the variants are acceptable.

As the accept mechanism is now usable for almost all situations, I
exposed the information of whether an input was rejected i.e. not
consumed by any coroutine that is waiting for input.